+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Upcoming
 
+-   `DecoderType` is now an alias for `$DecoderType` (to support both TypeScript and Flow
+    conventional naming)
+
 ## v1.21.0
 
 -   **New decoder** [`lazy()`](https://github.com/nvie/decoders#lazy): lazily-evaluated